journal_backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
aiosqlite==0.20.0
bcrypt==4.1.2
PyJWT==2.8.0
pydantic==2.5.3
python-multipart==0.0.7
